Implement BloomFilter._bit_offsets()
This method is the beating heart of our Bloom filter.  Given a value,
compute its corresponding bit offsets within the underlying string
representing our Bloom Filter.

Next, we'll make sure that all of the bits at those offsets are set to 1
when testing for membership; and we'll set all of the bits at those
offsets to 1 when inserting.