Cache bit array on BloomFilter object
Previously, we were loading the bit array from Memcache before every
`add()`, `update()`, `__contains__()`, and `__len__()` operation.

Instead, load the bit array once on `__init__()`.

The advantage is that we avoid chatter to Memcache, but the disadvantage
is that if another instance updates our bit array, we won't see the
changes.

I believe that the advantage far outweighs the disadvantage.  Our
typical use-case will be to instantiate a `BloomFilter`, then to use it
to pluck out recently consumed candidate links from a listing.  In this
use-case, we'll only have to hit Memcache once - on `__init__()`.  After
that, our tons of `__contains__()` calls will require no I/O.  And even
if our bit array gets updated by another instance in the meantime,
that's ok; we'll get those changes the next time that we instantiate our
`BloomFilter` (when we generate the next listing).