Implement Memcached RecentlyConsumed as set/deque
On the consumer side, we want a data structure similar to a
`collections.deque` (as we're appending a user's new consumed links to
the right and popping old consumed links from the left).  But on the
listing service side, we want a data structure similar to a `set` (as
we're testing many candidate links for membership within the user's
`RecentlyConsumed`).

Therefore, `RecentlyConsumed` (persisted in Memcache) implements a
subset of the `collections.deque` API and a subset of the `set` API.  As
such, please note that `RecentlyConsumed` silently ignores duplicate
values during `append()` and `extend()`.

Further, please note that `RecentlyConsumed` is neither thread-safe nor
safe to use across instances.  Please lock around your
`RecentlyConsumed` instances/accesses.  A good choice for such a lock is
`.memlock.MemLock`.