Implement Memcached RecentlyConsumed as set/deque (#18)
* Implement Memcached RecentlyConsumed as set/deque

On the consumer side, we want a data structure similar to a
`collections.deque` (as we're appending a user's new consumed links to
the right and popping old consumed links from the left).  But on the
listing service side, we want a data structure similar to a `set` (as
we're testing many candidate links for membership within the user's
`RecentlyConsumed`).

Therefore, `RecentlyConsumed` (persisted in Memcache) implements a
subset of the `collections.deque` API and a subset of the `set` API.  As
such, please note that `RecentlyConsumed` silently ignores duplicate
values during `append()` and `extend()`.

Further, please note that `RecentlyConsumed` is neither thread-safe nor
safe to use across instances.  Please lock around your
`RecentlyConsumed` instances/accesses.  A good choice for such a lock is
`.memlock.MemLock`.

* Namespace random keys to avoid accidental deletes

If you don't specify a key when instantiating any of these
Memcache-backed data structures, then the data structures assign a
random key that begins with `_RANDOM_KEY_PREFIX`.  Later on `__del__()`,
the data structures delete the Memcache key if it begins with
`_RANDOM_KEY_PREFIX`.

This commit namespaces the random keys to prevent `__del__()` from
accidentally deleting something from Memcache that you intended to keep.

* Unit test RecentlyConsumed / demonstrate usage